% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addgeodata.R
\name{addgeodata}
\alias{addgeodata}
\title{addgeodata}
\usage{
addgeodata(database = ISRaD_data, geodata_directory,
  geodata = "worldclim")
}
\arguments{
\item{database}{soilcarbon dataset object}

\item{geodata_directory}{directory where geospatial data is found}

\item{geodata}{character string for the geo spatial dataset you want to extract. default is "worldclim"}
}
\description{
adds geospatial data to soilcarbon database object and saves it to the soilcarbon package database
}
