% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTemplateFiles.R
\name{checkTemplateFiles}
\alias{checkTemplateFiles}
\title{Check ISRaD Template/Info files}
\usage{
checkTemplateFiles(outfile = "")
}
\arguments{
\item{outfile}{file to dump the output report. Defaults to an empty string that will print
to standard output.}
}
\value{
returns NULL
}
\description{
Check that the template information file and the template file match appropriately.
}
