% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTempetFiles.R
\name{checkTempletFiles}
\alias{checkTempletFiles}
\title{Check ISRaD Templet files}
\usage{
checkTempletFiles(outfile = "")
}
\arguments{
\item{outfile}{file to dump the output report. Defaults to an empty string that will print
to standard output.}
}
\value{
returns NULL
}
\description{
Check that the Templet information file and the templet file match appropreately.
}
\examples{
\dontrun{
checkTempletFiles()
}
}
