% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilations.R
\name{compilations}
\alias{compilations}
\title{compilations}
\usage{
compilations(file, compilation, write_out)
}
\arguments{
\item{file}{data file or directory}

\item{compilation}{character value indicating the data compilation. Acceptable values are "Yujie" and "Treat".}

\item{write_out}{T or F indicating whether or not to write out data tempate files in local directory}
}
\value{
a list with data organized in 'tabs' in template format
}
\description{
This function converts previous data complilations into the soilcarbon template. It can save individual files for each dataset as well as a single completed template file for the entire compilation.
}
