% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile}
\alias{compile}
\title{compile}
\usage{
compile(dataset_directory, write_report = F, write_out = F,
  return = NULL)
}
\arguments{
\item{dataset_directory}{directory where compeleted and QC passed soilcarbon datasets are stored}

\item{write_report}{T or F whether or not to write a log file of the compilation}

\item{write_out}{T or F whether or not to write the compiled database file as csv in dataset_directory}

\item{return}{parameter for whether compile function should return object to R environment. Default is NULL. Acceptable values are "flat" or "list" depending on the format you want to have the database returned in.}
}
\description{
adds dataset to soilcarbon database
}
