% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gravin.R
\docType{data}
\name{gravin}
\alias{gravin}
\title{Gravin dataset for delta delta calculation}
\format{dataframe}
\usage{
gravin
}
\description{
Data from Graven et al 2017 https://www.geosci-model-dev.net/10/4405/2017/gmd-10-4405-2017.pdf
}
\keyword{datasets}
