% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.build.R
\name{ISRaD.build}
\alias{ISRaD.build}
\title{ISRaD.build builds the database and updates objects in R package}
\usage{
ISRaD.build(ISRaD_directory = getwd(), geodata_directory)
}
\arguments{
\item{ISRaD_directory}{directory where the ISRaD package is found}

\item{geodata_directory}{directory where geospatial climate datasets are found. Necessary to create ISRaD_Extra}
}
\value{
runs QAQC on all datafiles, moves files that fail QAQC, updates ISRaD_Data, updates ISRaD_Extra
}
\description{
Wrapper function that combines tools for rapid deployment of R package data objects. 
Meant to be used by the maintainers/developers of ISRaD
}
\examples{
\dontrun{
ISRaD.build(ISRaD_directory="~/ISRaD/", geodata_directory="~/geospatial_datasets")
}
}
