% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.download.R
\name{ISRaD.download}
\alias{ISRaD.download}
\title{ISRaD.download}
\usage{
ISRaD.download(directory = NULL, extra = F)
}
\arguments{
\item{directory}{location for data to download. If NULL, data will go to current working directory.}

\item{extra}{T or F. If T, the ISRaD_extra object will be returned. If F, ISRaD_data will be returned. Default is F.}
}
\value{
ISRaD data object
}
\description{
ISRaD.download
}
