% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.Cstocks.R
\name{ISRaD.extra.Cstocks}
\alias{ISRaD.extra.Cstocks}
\title{ISRaD.extra.Cstocks}
\usage{
ISRaD.extra.Cstocks(database)
}
\arguments{
\item{database}{ISRaD dataset object.}
}
\description{
Calculates soil organic carbon stock
}
\details{
Function first fills bulk density and organic carbon concentration. SOC stocks can only be calculated if organic carbon concentration and bulk density data are available. SOC stock calculation is for the fine earth fraction (<2mm).
}
\references{
any references of literature or datasets relevant to understand the function. (remove this entire line if there are no references)
}
\author{
J. Beem-Miller
}
