% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.delta_delta.R
\name{ISRaD.extra.delta_delta}
\alias{ISRaD.extra.delta_delta}
\title{ISRaD.extra.delta_delta}
\usage{
ISRaD.extra.delta_delta(database)
}
\arguments{
\item{database}{ISRaD dataset object.}
}
\description{
: Calculates the difference between sample delta 14C and the atmosphere for the year of collection
}
\details{
: Creates new column for delta delta value. Observation year and profile coordinates must be filled (use ISRaD.extra.fill_dates, and ISRaD.extra.fill_coords fxs). Calls SoilR for atmospheric data. Atmospheric data are corrected for the mean northern or southern hemisphere value.
}
\references{
: Hua et al., 2013; Sierra et al., 2014
}
\author{
: J. Beem-Miller and C. Hicks-Pries
}
