% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.fill_expert.R
\name{ISRaD.extra.fill_expert}
\alias{ISRaD.extra.fill_expert}
\title{ISRaD.extra.fill_expert}
\usage{
ISRaD.extra.fill_expert(database)
}
\arguments{
\item{database}{ISRaD dataset object.}
}
\value{
returns ISRaD_data object with the lyr_xxx_fill_extra columns containing both original and filled data
}
\description{
: Fills in columns of expert-reviewed full data with real data where available, and calculates missing carbon stocks with filled data.
}
\details{
:
}
\examples{
\donttest {
ISRaD_full <- ISRaD.getdata(tempdir())
ISRaD.extra.fill_expert(ISRaD_full)
}
}
\references{
:
}
\author{
: Paul A. Levine
}
