% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.fill_fm.R
\name{ISRaD.extra.fill_fm}
\alias{ISRaD.extra.fill_fm}
\title{ISRaD.extra.fill_fm}
\usage{
ISRaD.extra.fill_fm(database)
}
\arguments{
\item{database}{ISRaD dataset object.}
}
\description{
Fills fraction modern values if missing and delta 14C is reported
}
\details{
: Warning: xxx_obs_date_y columns must be filled for this to work!
}
\references{
: Stuiver and Polach, 1977
}
\author{
: J. Beem-Miller & A. Hoyt
}
