% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.fill_soilorders.R
\name{ISRaD.extra.fill_soilorders}
\alias{ISRaD.extra.fill_soilorders}
\title{ISRaD.extra.fill_soilorders}
\usage{
ISRaD.extra.fill_soilorders(database)
}
\arguments{
\item{database}{ISRaD dataset object.}
}
\description{
extracts values from various geospatial climate datasets and adds to ISRaD_data object to create ISRaD_data_extra.
}
\details{
backfills pro_usda_soil_order based on USDA classificiations
}
