% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.flatten.R
\name{ISRaD.extra.flatten}
\alias{ISRaD.extra.flatten}
\title{ISRaD.extra.flatten}
\usage{
ISRaD.extra.flatten(database, table)
}
\arguments{
\item{database}{ISRaD dataset object: e.g. ISRaD_data, or ISRaD_extra}

\item{table}{ISRaD table of interest (flux, layer, interstitial, fraction, incubation)}
}
\description{
: Joins tables in ISRaD based on linking variables and returns "flat" dataframes
}
\details{
: ISRaD.extra.flatten generates flat files for user specfied ISRaD tables by joining higher level tables (metadata, site, profile, layer) to lower level tables (layer, fraction, incubation, flux, interstitial).
}
\references{
:
}
\author{
: J. Beem-Miller
}
