% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.geospatial.clay.R
\name{ISRaD.extra.geospatial.clay}
\alias{ISRaD.extra.geospatial.clay}
\title{ISRaD.extra.geospatial.clay}
\usage{
ISRaD.extra.geospatial.clay(database, geodata_soil_directory)
}
\arguments{
\item{database}{soilcarbon dataset object}

\item{geodata_soil_directory}{directory where geospatial soil clay data are found}
}
\value{
returns new ISRaD_extra object with extracted clay content up to 200 cm soil depth
}
\description{
Extracts modelled clay content from 250m resolution Soil Grids spatial products
}
\details{
Uses geographic coordinates of profiles (including those filled from site-level coordinates) to extract estimated (observations + machine learning predictions) clay content (kg/kg) from rasters at 0, 10, 30, 60, 100, and 200 cm soil depth. Rasters provided by SoilGrids (doi: 10.5281/zenodo.2525663).
}
