% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.geospatial.soil.R
\name{ISRaD.extra.geospatial.soil}
\alias{ISRaD.extra.geospatial.soil}
\title{ISRaD.extra.geospatial.soil}
\usage{
ISRaD.extra.geospatial.soil(database, geodata_directory)
}
\arguments{
\item{database}{ISRaD dataset object.}

\item{geodata_directory}{directory where geospatial climate datasets are found.}
}
\description{
brief summary description of function
}
\details{
very detailed description of function, especially if it involves the creation of new columns
}
\references{
any references of literature or datasets relevant to understand the function. (remove this entire line if there are no references)
}
\author{
your name
}
