% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.fill.R
\name{ISRaD.fill}
\alias{ISRaD.fill}
\title{ISRaD.fill}
\usage{
ISRaD.fill(database = ISRaD_data, geodata_directory)
}
\arguments{
\item{database}{soilcarbon dataset object}

\item{geodata_directory}{directory where geospatial data is found}
}
\description{
fills in transformed data, or empty NA values where possible, and adds geospatial data to soilcarbon database object
}
