% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.getdata.R
\name{ISRaD.getdata}
\alias{ISRaD.getdata}
\title{ISRaD.getdata}
\usage{
ISRaD.getdata(directory = getwd(), dataset = "full", extra = F)
}
\arguments{
\item{directory}{location of ISRaD_database_files folder. If not found, it will be download. The default is the current working directory.}

\item{dataset}{Specify which data you want. Options are c("full", flux","interstitial","incubation","fraction","layer")}

\item{extra}{T or F. If T, the ISRaD_extra object will be returned. If F, ISRaD_data will be returned. Default is F.}
}
\value{
ISRaD data object
}
\description{
ISRaD.getdata
}
