% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.save.excel.R
\name{ISRaD.save.xlsx}
\alias{ISRaD.save.xlsx}
\title{ISRaD.save.xlsx}
\usage{
ISRaD.save.xlsx(database, template_file, outfile)
}
\arguments{
\item{database}{ISRaD dataset object.}

\item{template_file}{path and name of template file to use.}

\item{outfile}{path and name to save the excel file}
}
\description{
saves data object as xlsx file in ISRaD template format
}
\author{
J Grey Monroe
}
