% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD_data.R
\docType{data}
\name{ISRaD_data}
\alias{ISRaD_data}
\title{ISRaD database object}
\format{A list of data frames. The names of the data frames and their columns reflect the
 structure of the ISRaD data master template.}
\usage{
ISRaD_data
}
\description{
Complete database object compiled for ISRaD
diamonds.
}
\keyword{datasets}
