% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD_data_filled.R
\docType{data}
\name{ISRaD_data_filled}
\alias{ISRaD_data_filled}
\title{ISRaD extra database object}
\format{A list of data frames. The names of the data frames and their columns reflect the
 structure of the ISRaD data master template. Additional columns have been added and certain 
 variables have been filled in when possible.}
\usage{
ISRaD_data_filled
}
\description{
Complete database object compiled for ISRaD. Includes extra variables calculated using ISRaD.extra function.
diamonds.
}
\keyword{datasets}
