% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Treat2016.R
\name{read_Treat2016}
\alias{read_Treat2016}
\title{Read in data for Treat 2016.}
\usage{
read_Treat2016(dowloadDir = "temp")
}
\arguments{
\item{dowloadDir}{}
}
\value{

}
\description{
Currently doesn't work and is under development
}
