% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{readYujie}
\alias{readYujie}
\title{readYujie}
\usage{
readYujie(Yujie_file)
}
\arguments{
\item{Yujie_file}{directory to Yuije data file}
}
\value{
a list with data organized in 'tabs'
}
\description{
This function converts data compiled by Yuije (in csv) into a format compatible with the soilcarbon template
}
