% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_count_data.R
\name{rep_count_data}
\alias{rep_count_data}
\title{rep_count_data}
\usage{
rep_count_data(database = NULL)
}
\arguments{
\item{database}{ISRaD data object}
}
\description{
generate a count of observations for each level of the database
}
