% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_entry_stats.R
\name{rep_entry_stats}
\alias{rep_entry_stats}
\title{rep_entry_stats}
\usage{
rep_entry_stats(database = NULL)
}
\arguments{
\item{database}{ISRaD data object}
}
\description{
generate report of entry statistics
}
