% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_frc_data.R
\name{rep_frc_data}
\alias{rep_frc_data}
\title{rep_frc_data}
\usage{
rep_frc_data(database = NULL)
}
\arguments{
\item{database}{ISRaD data object}
}
\description{
generate a count of fractionation observations including scheme and property
}
