% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_site_map.R
\name{rep_site_map}
\alias{rep_site_map}
\title{rep_site_map}
\usage{
rep_site_map(database = NULL)
}
\arguments{
\item{database}{ISRaD data object}
}
\description{
generate a world map with site locations plotted
}
