% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{reports}
\alias{reports}
\title{reports}
\usage{
reports(database = NULL, report)
}
\arguments{
\item{database}{ISRaD data object}

\item{report}{Parameter to define which type of report you want. Options are, "entry_stats", "flattened", "fraction"...}
}
\description{
generate reports of ISRaD data
}
