% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.delta_delta.R
\name{ISRaD.extra.delta_delta}
\alias{ISRaD.extra.delta_delta}
\title{ISRaD.extra.delta_delta}
\usage{
ISRaD.extra.delta_delta(database)
}
\arguments{
\item{database}{ISRaD dataset object.}
}
\value{
returns ISRaD_data object with new delta delta columns in relevant tables
}
\description{
Calculates the difference between sample delta 14C and the atmosphere for the year of collection (delta-delta)
}
\details{
Creates new column for delta-delta value. Observation year and profile coordinates must be filled (use ISRaD.extra.fill_dates, and ISRaD.extra.fill_coords fxs). The relevant atmospheric d14C data (northern or southern hemisphere, or tropics) are determined by profile coordinates.
}
\references{
Graven et al. 2017 https://www.geosci-model-dev.net/10/4405/2017/gmd-10-4405-2017.pdf
}
\author{
J. Beem-Miller and C. Hicks-Pries
}
