% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.delta_delta.R
\name{ISRaD.extra.delta_delta}
\alias{ISRaD.extra.delta_delta}
\title{ISRaD.extra.delta_delta}
\usage{
ISRaD.extra.delta_delta(database)
}
\arguments{
\item{database}{ISRaD dataset object.}
}
\description{
brief summary description of function
}
\details{
very detailed description of function, especially if it involves the creation of new columns
}
\references{
any references of literature or datasets relevant to understand the function. (remove this entire line if there are no references)
}
\author{
your name
}
