% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.fill_14c.R
\name{ISRaD.extra.fill_14c}
\alias{ISRaD.extra.fill_14c}
\title{ISRaD.extra.fill_14c}
\usage{
ISRaD.extra.fill_14c(database)
}
\arguments{
\item{database}{ISRaD dataset object.}
}
\description{
: Fills delta 14C if not reported and fraction modern and
}
\details{
: Warning: xxx_obs_date_y columns must be filled for this to work!
}
\references{
: Stuiver and Polach, 1977
}
\author{
: J. Beem-Miller & A. Hoyt
}
