% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.fill_dates.R
\name{ISRaD.extra.fill_dates}
\alias{ISRaD.extra.fill_dates}
\title{ISRaD.extra.fill_dates}
\usage{
ISRaD.extra.fill_dates(database)
}
\arguments{
\item{database}{ISRaD dataset object.}
}
\value{
returns ISRaD_data object with filled obs_date_y columns
}
\description{
Fills frc_obs_date_y and inc_obs_date_y columns from lyr_obs_date_y if not reported.
}
\details{
This function must be run prior to the ISRaD.extra.fill_14c, ISRaD.extra.fill_fm, and ISRaD.extra.delta_delta for the layer and fraction tables.
}
\examples{
\donttest {
ISRaD_full <- ISRaD.getdata(tempdir())
ISRaD.extra.fill_dates(ISRaD_full)
}
}
