% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.extra.R
\name{ISRaD.extra}
\alias{ISRaD.extra}
\title{ISRaD.extra}
\usage{
ISRaD.extra(database, geodata_clim_directory, geodata_soil_directory,
  geodata_pet_directory)
}
\arguments{
\item{database}{soilcarbon dataset object}

\item{geodata_clim_directory}{directory where geospatial climate data are found}

\item{geodata_soil_directory}{directory where geospatial soil data are found}

\item{geodata_pet_directory}{directory where geospatial pet data are found}
}
\value{
returns new ISRaD_extra object with derived, transformed, and filled columns
}
\description{
Fills in transformed and geospatial data where possible, generatating an enhanced version of ISRaD.
}
\details{
Fills fraction modern, delta 14C, delta-delta values, profile coordinates, and SOC stocks frmo entered data; fills soil and climatic data from external geospatial data products
}
