% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRaD.getdata.R
\name{ISRaD.getdata}
\alias{ISRaD.getdata}
\title{ISRaD.getdata}
\usage{
ISRaD.getdata(directory = getwd(), extra = F, flat = F, tab = NULL)
}
\arguments{
\item{directory}{location of ISRaD_database_files folder. If not found, it will be download. The default is the current working directory.}

\item{extra}{T or F. If T, the ISRaD_extra object will be returned. If F, ISRaD_data will be returned. Default is F.}

\item{flat}{T or F. If T, the function will return the flattened data object.}

\item{tab}{if flat == T, you must specify which flattened file you want. Options are c("flux","interstitial","incubation","fraction","layer")}
}
\value{
ISRaD data object
}
\description{
ISRaD.getdata
}
