% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAQC.R
\name{QAQC}
\alias{QAQC}
\title{QAQC}
\usage{
QAQC(file, writeQCreport = F, outfile_QAQC = "", summaryStats = T,
  dataReport = F, checkdoi = T, verbose = T)
}
\arguments{
\item{file}{directory to data file}

\item{writeQCreport}{if TRUE, a text report of the QC output will be written to the outfile. Default is FALSE}

\item{outfile_QAQC}{filename of the output file if writeQCreport==TRUE. Default is NULL, and the outfile will be written to the directory where the dataset is stored, and named by the dataset being checked.}

\item{summaryStats}{prints summary statistics. Default is TRUE}

\item{dataReport}{prints list structure of database. Default is FALSE}

\item{checkdoi}{set to F if you do not want the QAQC check to validate doi numbers}

\item{verbose}{set to TRUE to print results of function}
}
\description{
Check the imported soil carbon dataset for formatting and entry errors
}
\details{
This function is also called by the online QAQC tool available at the ISRaD website <https://international-soil-radiocarbon-database.ocpu.io/ISRaD/www/>.
}
