% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Treat2016.R
\name{read_Treat2016}
\alias{read_Treat2016}
\title{Read in data for Treat 2016.}
\usage{
read_Treat2016(dowloadDir = "temp")
}
\arguments{
\item{dowloadDir}{directory where data files will be downloaded}
}
\value{
writes out files for individual data objects
}
\description{
Currently doesn't work and is under development
}
\examples{
\dontrun{
read_Treat2016()
compile("~/Dropbox/USGS/ISRaD_Data/Compilations/Treat/converted/")
}
}
