% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Treat2016.R
\name{read_Treat2016}
\alias{read_Treat2016}
\title{Read in data for Treat 2016.}
\usage{
read_Treat2016(download = T, downloadDir = "temp",
  convertedDir = "~/Dropbox/USGS/ISRaD_data/Compilations/Treat/converted/",
  dois_file = "~/Dropbox/USGS/ISRaD_data/Compilations/Treat/dois.csv")
}
\arguments{
\item{download}{boolean, if T the Treat datasets will be downloaded from pangea. Otherwise, they files in downloadDir will be used.}

\item{downloadDir}{directory where data files will be downloaded}

\item{convertedDir}{directory where data files that are converted to ISRaD template will be saved}

\item{dois_file}{file with doi numbers}
}
\value{
writes out files for individual data objects
}
\description{
Currently doesn't work and is under development
}
